// src/advertising/dto/credentials.ads.dto.ts
export interface CredentialsAdsDto {
  seller_id: string;
  marketplace_id: string;
  country_code: string;
  currency_code: string;
  daily_budget: number;
  time_zone: string;
  return_url: string;
  access_token: string;
  refresh_token: string;
  seller_name: string;
  client_id: string;
  client_secret: string;
  token_type: string;
  access_token_expiry: number;
  profile_id: number;
  region: string;
  access_token_exprity: number;
}